+++ volmdlr/edges.py
@@ -21,8 +21,8 @@ from matplotlib import __version__ as _mpl_version
 from mpl_toolkits.mplot3d import Axes3D
 from packaging import version
 
-import dessia_common as dc
 import plot_data.core as plot_data
+import dessia_common.core as dc
 import volmdlr.core_compiled
 import volmdlr.core
 import volmdlr.geometry
@@ -1530,8 +1530,14 @@ class LineSegment2D(LineSegment):
         return circle1, circle2
 
     def infinite_primitive(self, offset):
+<<<<<<< HEAD
         n = self.normal_vector()
         offset_point_1 = self.start + offset * n
+=======
+        n = self.unit_normal_vector()
+        offset_point_1 = self.start + offset * \
+            n
+>>>>>>> master
 
         offset_point_2 = self.end + offset * n